fix: move get_jpb_output func to utils file
