+++ test/e2e/conftest.py
@@ -12,15 +12,11 @@ import tempfile
 from dataclasses import dataclass, field, InitVar
 from typing import Generator, Type
 from contextlib import contextmanager
-from deadline.job_attachments._aws.deadline import get_queue
-from deadline.job_attachments import download
 
 from deadline_test_fixtures import (
     DeadlineWorker,
     DeadlineWorkerConfiguration,
     DockerContainerWorker,
-    DeadlineClient,
-    Job,
     Farm,
     Fleet,
     Queue,
@@ -373,29 +369,3 @@ def operating_system(request) -> OperatingSystem:
         return OperatingSystem(name="AL2023")
     else:
         return OperatingSystem(name="WIN2022")
-
-
-def get_job_output(
-    job: Job, deadline_client: DeadlineClient, deadline_resources: DeadlineResources
-) -> dict[str, list[str]]:
-    job.wait_until_complete(client=deadline_client, max_retries=20)
-
-    job_attachment_settings = get_queue(
-        farm_id=deadline_resources.farm.id,
-        queue_id=deadline_resources.queue_a.id,
-    ).jobAttachmentSettings
-
-    assert job_attachment_settings is not None
-
-    job_output_downloader = download.OutputDownloader(
-        s3_settings=job_attachment_settings,
-        farm_id=deadline_resources.farm.id,
-        queue_id=deadline_resources.queue_a.id,
-        job_id=job.id,
-        step_id=None,
-        task_id=None,
-    )
-    output_paths_by_root = job_output_downloader.get_output_paths_by_root()
-    job_output_downloader.download_job_output()
-
-    return output_paths_by_root
